qa/l7/requirements.txt
# Основные зависимости для мобильного тестирования
Appium-Python-Client==5.2.4
selenium==4.36.0

# Дополнительные зависимости
requests==2.31.0
Pillow==10.0.1

# Зависимости Selenium
attrs==25.3.0
certifi==2025.10.5
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
